requirements.txt
pandas
psycopg2-binary
dbt-postgres
